+++ src/main/java/roomescape/user/repository/reservation/InMemoryReservationRepository.java
@@ -1,4 +1,4 @@
-package roomescape.user.repository;
+package roomescape.user.repository.reservation;
 
 import java.util.ArrayList;
 import java.util.List;
